Concerted intermediate language to 64 bit (prevents segmentation faults). Added MiniExample and used it as the default for the compiler because ArrSum is still not compiled properly.

+++ runtime.h
@@ -1,10 +1,10 @@
 #ifndef _runtime_h
 #define _runtime_h
 
-int32_t L_halloc(int32_t);
-int32_t L_println_int(int32_t);
-int32_t L_print_char(int32_t);
-int32_t L_raise(int32_t);
+int64_t L_halloc(int64_t);
+int64_t L_println_int(int64_t);
+int64_t L_print_char(int64_t);
+int64_t L_raise(int64_t);
 
 
 #endif
